Changes in interface, methods names and places and code style
The interface of the application have been modified to show the current webpage name with a different button color of the main menu. Also the files listed have been changed according of rubocop code style.
