+++ app/controllers/application_controller.rb
@@ -12,7 +12,7 @@ class ApplicationController < ActionController::Base
 
   def after_sign_in_path_for(resource)
     if user_signed_in?
-      stored_location_for(resource) || monthly_sales_path
+      stored_location_for(resource) || overview_months_path
     else
       root_path
     end
