+++ app/controllers/concerns/database_searchs.rb
@@ -1,131 +1,7 @@
 module DatabaseSearchs
-  def fetch_last_report
-    Report.where(month_numb: fetch_last_month, year: fetch_last_year)[0]
-  end
-
-  def fetch_last_year
-    Report.distinct.order('year').pluck(:year)[-1]
-  end
-
-  def fetch_last_month
-    Report.distinct.where(year: Time.current.year)
-          .order('month_numb').pluck(:month_numb)[-1]
-  end
-
-  def fetch_report(month, year)
-    Report.where(month: month, year: year.to_i)[0]
-  rescue StandardError
-    nil
-  end
-
-  def prepare_fetch_reports_by_month_range(report)
-    @i = 1
-    @first_month = report.month_numb
-    @first_year = report.year
-  end
-
-  def fetch_reports_by_month_range(current_report, quantity = 1)
-    prepare_fetch_reports_by_month_range(current_report)
-
-    while @i < quantity
-      @first_month -= 1
-      if @first_month.zero?
-        @last_to_first = 12
-        @last_year -= 1
-      end
-      @i += 1
-    end
-
-    Report.where(month_numb: @first_month, year: @first_year)[0]
-  end
-
-  def fetch_report_by_next_month(current_report)
-    if current_report.month_numb + 1 > 12
-      fetch_report('January', (current_report.year + 1))
-    else
-      Report.where(month_numb: current_report.month_numb + 1,
-                   year: current_report.year)[0]
-    end
-  rescue StandardError
-    nil
-  end
-
-  def fetch_all_years
-    Report.distinct.order('year').pluck(:year).reverse
-  end
-
-  def fetch_reports_by_year(year)
-    Report.where(year: year).order('month_numb').reverse
-  end
-
-  def fetch_reports_by_current_year
-    Report.where(year: fetch_last_year).order('month_numb').reverse
-  end
-
-  def fetch_report_by_unique_years(same_year = nil)
-    list = Report.distinct.pluck(:year).reverse
-    list.delete(same_year)
-    list.unshift(same_year)
-  end
-
-  def fetch_report_by_id(id)
-    Report.find_by(id: id)
-  end
-
-  #--------------------------------------------------------------------------##
-  #--------------------------------------------------------------------------##
-  #--------------------------------------------------------------------------##
-
-  def fetch_contract_by_report_id(report_id)
-    Contract.where(report_id: report_id).order('day')
-  end
-
-  def destroy_contract_by_id(contract_id)
-    Contract.find_by(id: contract_id).destroy
-  end
-
-  def fetch_contracts_by_day_report_id(day, report_id)
-    Contract.where(day: day, report_id: report_id)
-  end
-
-  def fetch_contracts_by_user_report_id(user_id, report_id)
-    Contract.where(user_id: user_id, report_id: report_id).order('day')
-  end
-
-  def fetch_contract_sum(user_id, report_id)
-    values = 0
-    contracts = fetch_contracts_by_user_report_id(user_id, report_id)
-    contracts.each { |contract| values += contract.value }
-    values
-  end
-
-  #--------------------------------------------------------------------------##
-  #--------------------------------------------------------------------------##
-  #--------------------------------------------------------------------------##
-
-  def fetch_username_by_id(user_id)
-    User.find_by(id: user_id).full_name
-  end
-
-  def fetch_id_by_username(full_name)
-    User.find_by(full_name: full_name).id
-  end
-
-  def fetch_username_by_priority
-    User.where('priority BETWEEN ? AND ?', 1, 2).order('full_name')
-  end
-
-  def fetch_user_by_priority(priority)
-    User.where(priority: priority).order('full_name')
-  end
-
-  def fetch_username_by_salesman_priority
-    fetch_user_by_priority(1).collect(&:full_name)
-  end
-
-  #--------------------------------------------------------------------------##
-  #--------------------------------------------------------------------------##
-  #--------------------------------------------------------------------------##
+  include ReportSearchs
+  include ContractSearchs
+  include UserSearchs
 
   def month_days
     Time.days_in_month(@current_report.month_numb)
