+++ app/views/dashboard/monthly_sales.html.slim
@@ -4,6 +4,8 @@ body onload="generateGraphicPlot(#{@report_points}, #{@contract_points}, #{@day_
   .row
     .panel.panel-default style="margin-top: -30px;"
       .panel-heading
+        h3.h3.form-group.text-center
+          = "Monthly Sales"
         .form-inline
           h3.h3.form-group
             = "Name:"
