+++ app/views/dashboard/overview_months.html.slim
@@ -5,6 +5,8 @@ body onload="generateOverviewPlot(#{@goal_points}, #{@sum_points}, #{@month_text
     .row
       .panel.panel-default
         .panel-heading
+          h3.h3.form-group.text-center.m-b
+            = "Overview Month"
           .form-inline
             = simple_form_for @report, method: :get, url: "/overview_months", html: { class: "form-group" } do |f|
               h3.h3.form-group.m-r
@@ -18,6 +20,7 @@ body onload="generateOverviewPlot(#{@goal_points}, #{@sum_points}, #{@month_text
                 = f.input :year, label: false, :collection => @last_list, :prompt => false, input_html: { class: "form-control", id: "monthBox", align: "central"}
 
               = f.button :submit, "Search", class: "btn btn-primary btn-addon btn-lg margin m-l"
+
         .panel-body style="font-size: 25px; font-weight: bold;"
           div id="placeholderO" style="height:450px;width:1250px;"
           = simple_form_for @report, method: :get, url: "/monthly_sales", html: { class: "form-group", id: "formSearch" } do |f|
