+++ app/views/dashboard/overview_reports.html.slim
@@ -1,10 +1,12 @@
 head
   title Sales Up - Overview Report
-body onload="generateOverviewRecordPlot(#{@salesman_points}, #{fetch_username_by_salesman_priority}, #{@month_text})"
+body onload="generateOverviewRecordPlot(#{@salesman_points}, #{@usernames}, #{@month_text})"
   div style="margin-top: -30px;"
     .row
       .panel.panel-default
         .panel-heading
+          h3.h3.form-group.text-center.m-b
+            = "Overview Report"
           .form-inline
             = simple_form_for @report, method: :get, url: "/overview_reports", html: { class: "form-group" } do |f|
               h3.h3.form-group.m-r
@@ -17,7 +19,16 @@ body onload="generateOverviewRecordPlot(#{@salesman_points}, #{fetch_username_by
                 - @last_list.delete(@first_list[0])
                 = f.input :year, label: false, :collection => @last_list, :prompt => false, input_html: { class: "form-control", id: "monthBox", align: "central"}
 
+              h3.h3.form-group.m-r
+                = "Compare: "
+                = f.input :report_name, label: false, :collection => [["All Salesman", "All"]] + fetch_username_by_salesman_priority.collect { |username| [ username, username] }, :prompt => false, input_html: { class: "form-control", id: "monthBox" }
+
+              h3.h3.form-group.m-r
+                = "With: "
+                = f.input :goal, label: false, :collection => [["Contract Sum", "CS"], ["Contracts Closed", "CC"], ["Goal's Percentage", "CP"]], :prompt => false, input_html: { class: "form-control", id: "monthBox" }
+
               = f.button :submit, "Search", class: "btn btn-primary btn-addon btn-lg margin m-l"
+
         .panel-body style="font-size: 25px; font-weight: bold;"
           div id="placeholderOR" style="height:450px;width:1250px;"
